+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Unreleased
 ----------
+###Added
+- Added `Templates` class.
+
 ### Changed
 - Updated `README.md` docs.